imported Grid from CS_Grid_maker #34
