+++ brix/__init__.py
@@ -2,4 +2,4 @@
 from .classes import *
 from .functions import *
 from .test_tools import User, spin_users, stop_users, list_users
-from .grid_maker import Grid, commit_grid, edit_types
+from .grid_maker import Grid
