+++ requirements.txt
@@ -6,4 +6,4 @@ shapely>=1.6.2
 geopandas>=0.6.2
 geopy>=1.20.0
 networkx>=2.4
-matplotlib>=3.0.3
+pyproj==1.9.5.1
